Resistances implemented from 'Resistance Board' abstract class. Added new class 'State Bar' for health, mana and items condition properties. Using float type instead of byte.

+++ Models/Items/Usable/Healing/HealingItem.cs
@@ -5,9 +5,9 @@ namespace Models.Items.Usable
     public abstract class HealingItem : UsableItem
     {
 
-        public byte Recovery { get; }
+        public float Recovery { get; }
 
-        public HealingItem(string name, byte recovery) : base(name)
+        public HealingItem(string name, float recovery) : base(name)
         {
             Recovery = recovery;
         }
